✨ 그룹 문제집 활동 상세 페이지 api 연동 : 그룹 문제집 정보 조회
